rewrite jiggle/letter logic + create demo
In this commit, I improved the letter export with a custom setter that
will make the changes apparent in the editor without having to manually
set the texture/frame.

I also added a jiggle export, which also has a custom setter. This
allows for toggling the jiggling of the letter blocks. If it is set to
false, it will correctly reorient the letter to 0 radians.

Finally, I moved the letters from the CanvasGroup to the TileMap (it just
makes sense), and set up a demo of some letters not moving. They also
can have the letter changed or the jiggle toggled while the debug mode
window is running, through the Remote panel!