add ExtLetterManager and add to scene trees
ExtLetterManager will be in charge of manipulating letters in the game
area. It will not handle logic or know anything about what letter any of
the letters are. It will only know how to handle clicks on the letters
and will move them as requested.

Currently, ExtLetterManager only has the init logic implemented.