add FixedVec<T> to letter_manager module
This newtype will help me express using strong typing what I had
attempted to express using errors. I want a Vec that can't be resized
after allocation to represent nodes and positions that will remain fixed
until the game is over, hence FixedVec.

Under the hood, it uses a custom Deref/DerefMut implementation to skip
the Vec methods and go straight to the underlying slice. This prevents
usage of typical Vec methods such as push() and pop(), but allows the
functionality of creating a dynamic-sized array on the heap. Why not use
Box<[T]>? I want the option of growing the slice at runtime, which a Box
does not allow. into_boxed_slice() is not free:

[Link](https://www.reddit.com/r/rust/comments/19b7a0v/comment/kipy76b/)

Maybe there is a library that does this, maybe even better. But why
would I depend on a library for such simple functionality? In case I
need it for other projects, I have implemented all the usual traits.