added ability to move letters back and forth
This is the core of ExtLetterManager's functionality: it handles moving
the letters! Eventually, I would like to add animations to this. I will
also need ExtLetterManager to notify that letters have changed, and
provide the current status of the words constructed to the game node.
This will ensure we have separation of concerns, and will make the game
much more maintainable.