change Words.get() to accept StringName
While it does make sense for the words module to handle converting the
letters to words, it is easier for interoperation for Words.get() to
instead accept StringName. This is because StringName both doesn't use
a raw pointer, so can be deserialized, and can be passed through Godot
API's. Maybe at some point I will move this back as a standalone
function of some sort in this module.