change Words implementation + add validation tests
Deserializing directly to StringName (maybe) has a slight performance
benefit, but this is outweighed by the fact that this representation
prevents unit testing. This is because StringName requires the Godot
engine to be available, which it isn't in the context of library testing.
So instead we use Rust Strings.

As a result of this change, testing if the expected files exist is
possible, and implemented in this commit.